refactor: extract demo initialization helper
